Don't wait for clients to be present inorder to read program I/O
Previously the underlying application would block on I/O because
we would wait to read it until a client is connected to display
the information. This prevented long running tasks from making
progress. However with the new approach output data of programs
which terminate while no client is connected is lost.