Revert "Set sticky bit on socket"
This reverts commit 15ae398ea51f9adb282d7cf93a998a8ea401fd21.

Setting the sticky bit on a file is not really portable and
causes more problems than it solves.

The reason why this was done in the first place, is that the
XDG Base Directory Specification states that files with the
sticky bit set should not be touched by periodic cleanup
procedures.

Close #13