Do not implicilty enable passthrough mode if no session name is given
Even when no session name was provided the passthrough flag was still
implicitly enabled in case stdin was not a terminal. This behavior
prevented the session list from being printed i.e.

  abduco < /dev/null

would not work as expected. Only enable implicit passthrough if stdin
is not a terminal *and* a session name is given.

Fix #34