fix bug where attaching to dead session won't give underlying exit code
If dvtm dies, and you run `:|abduco -a session`, abduco will often exit
without giving you the exit code, and without removing the socket from
~/.abduco.

We resolve this by first sending the client either an EXIT or an empty
CONTENT packet.

Only after the client recieves it, we set the socket to non-blocking.

Fixes: https://github.com/martanne/abduco/issues/44